@@ -5,12 +5,12 @@ module DataScraper
 #  require 'pry-debugger'
 
   include CommonDefs
-  YEAR = Time.new.year 
+  YEAR = Time.new.year
 
-  # The above include covers the following three definitions  
+  # The above include covers the following three definitions
   # MILLION = 1000000
   # BILLION = 1000000000
-  # YEAR = Time.new.year 
+  # YEAR = Time.new.year
 
   def scrape_data(mark = "0")
 
@@ -21,52 +21,52 @@ module DataScraper
       puts "Skipping downloading data for #{ticker} - not going to download"
       return true
     end
-    
+
     # Check if earnings data look strange, if so delete them
     earnings2012 = eps.select{|e| e.year == 2012 }.first
     earnings2011 = eps.select{|e| e.year == 2011 }.first
     condition = earnings2012 && earnings2011 && earnings2012.revenue == earnings2011.revenue && earnings2012.net_income == earnings2011.net_income
-    if condition        
+    if condition
       eps.map{ |s| s.delete }
       self.reload
       puts "[NOTE] earnings deleted for #{ticker} beacuse they apear to be erroneous"
     end
-      
+
     # retriev:
     # From: gurufocus.com
     get_data_from_gurufocus # Retrievs: ta,tl, ca, cl, ltd, nta, bv
                             # Revenue, income, and eps
-    get_numshares           # numshares - Not needed for first 1000 up to NU 
+    get_numshares           # numshares - Not needed for first 1000 up to NU
     update_price
     get_dividends
 
 #    update_current_data #This works, but takes a long time to run
 # since it makes multiple requests, - shorten number of requests!
     get_eps
-    
+
     # mark stock as updated
-    update_attributes( :mark => markn) 
+    update_attributes( :mark => markn)
   end
 
   def yearly_update
-    
+
     get_balance_sheets
     get_income
     get_book_value
     get_sales
-    
+
     # get_dividends(1900) # get dividends as far back as possible
-    
+
     get_last_year_eps_msn
     get_historic_eps(10) # get earnings for last year
     # get_eps # ttmeps
-    
+
     update_current_data # ttm_eps, sales, div_yield
-          
+
     get_numshares
-    
+
     update_price
-    
+
   end
 
   def divs
@@ -86,29 +86,29 @@ module DataScraper
   def get_quarterly
     # get last year earnings
     l_year = latest_eps.year
-  
+
     # get which quarters are the last 4
     fp = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Fiscal Period" }
     fp = fp.xpath('./td') if fp
-       
+
     if fp.nil?
       puts "--------------------------------------Cannot get info for #{ticker}"
       return false
-    end   
+    end
     # Find last year by counting 'td's up to "TMM"
     years_available = 0 # Some stocks may not have 10 years worth of data
     for i in 1..fp.size
-      if fp[i].nil? || !fp[i].text.match("TTM").nil? 
+      if fp[i].nil? || !fp[i].text.match("TTM").nil?
         break
       end
       years_available = i
     end
-    
+
     puts "Counted #{years_available} years of available data for #{ticker}"
-          
+
     update_year = 1 # Some stocks may not be updated for 2012 yet
     update_year = 0 if fp[years_available].text.last == "2"
-    
+
 
 
 
@@ -119,56 +119,56 @@ module DataScraper
       puts "Could not get quarterly finantial data from gurufocus.com"
       return false
     end
-   
+
     # Get last 4 quarters quarterly data
     # Check first if all 4 quarters are available?
     (1..4).each do |i|
-      if fp[i].nil? || !fp[i].text.match("TTM").nil? 
+      if fp[i].nil? || !fp[i].text.match("TTM").nil?
         break
       end
       years_available = i
     end
-    
+
     puts "Counted #{years_available} years of available data for #{ticker}"
-          
+
     update_year = 1 # Some stocks may not be updated for 2012 yet
     update_year = 0 if fp[years_available].text.last == "2"
-    
+
     # A boolean to test if current asset values are available
     using_current_data = true
-    
+
     # Scrape data from doc
     # Current Assets
     ca = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Current Assets" }
     if ca
-      ca = ca.xpath('./td') 
+      ca = ca.xpath('./td')
     else
       using_current_data = false
     end
-    
+
     ta = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Assets" }
     ta = ta.xpath('./td') if ta
-    
+
     cl = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Current Liabilities" }
     if cl
-      cl = cl.xpath('./td') 
+      cl = cl.xpath('./td')
     else
       using_current_data = false
     end
-        
+
     tl = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Liabilities" }
     tl = tl.xpath('./td') if tl
-        
+
     # Debt, book value, net tangible assets
     ltd = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Long-Term Debt" }
     ltd = ltd.xpath('./td') if ltd
-    
+
     bv = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Equity" }
     bv = bv.xpath('./td') if bv
-    
+
     ocs = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Other Current Assets" }
     ocs = ocs.xpath('./td') if ocs
-    
+
     # Create balance sheet for 10 years
     (1..years_available).each do |i|
       cas = ""
@@ -181,9 +181,9 @@ module DataScraper
           ntas = (( clean_string(ca[i].text).to_f - clean_string(ocs[i].text).to_f - clean_string(cl[i].text).to_f ).round * MILLION ).to_s
         else
           ntas = cas
-        end 
+        end
       end
-      
+
       # Some trusts don't have liabilities
       tler = ""
       tler = (clean_string(tl[i].text).to_f.round * MILLION).to_s if tl
@@ -200,18 +200,18 @@ module DataScraper
                             :long_term_debt => der,
                             :net_tangible_assets => ntas,
                             :book_value => bver,
-                            :quarter => q)         
+                            :quarter => q)
         puts "Got bs data for #{ticker}, year: #{bs.year}, ta = #{bs.total_assets}" if !bs.id.nil?
     end
-    
-    update_attributes( :has_currant_ratio => using_current_data)    
+
+    update_attributes( :has_currant_ratio => using_current_data)
 
   end # end quarterly ---------------------------------------------
 
 
 
   def get_stock_price
-    price = get_price_from_msn  
+    price = get_price_from_msn
     price = get_price_from_google if price.nil?
     price                                                                         end
 
@@ -263,7 +263,7 @@ module DataScraper
 
 
   # Price scrapers -----------------------------------------------------------
- 
+
   def get_price_from_google
     url = "http://www.google.com/finance?q=#{ticker}"
     doc = open_url_or_nil(url)
@@ -276,12 +276,13 @@ module DataScraper
   end
 
   def get_price_from_msn
-    url = "http://investing.money.msn.com/investments/stock-price?symbol=us%3a#{ticker}"
+    url = "http://www.msn.com/en-us/money/stockdetails/fi-126.1.#{ticker}"
+
     doc = open_url_or_nil(url)
-    price = doc.css('#quickquoteb') if doc
+    price = doc.css('.precurrentvalue') if doc
     if price
       begin
-        price = clean_string( price.children[3].children[1].text ).to_f 
+        price = clean_string( price.text ).to_f
       rescue
       end
     end
@@ -290,11 +291,11 @@ module DataScraper
 
 # Revenue and Incom scrapper from msn --------------------------------------
 def get_revenue_income_msn
-    year = Time.new.year 
+    year = Time.new.year
     puts "Getting Revenue and Income for #{ticker}"
-   
+
     url = "http://investing.money.msn.com/investments/stock-income-statement/?symbol=US%3a#{ticker}"
-    
+
     doc = open_url_or_nil(url)
 
     if doc
@@ -309,11 +310,11 @@ def get_revenue_income_msn
 
     # Msn gives numbers in millions, so we will multiply by 1000000
     if tr && ni
-      
+
       (1..5).each do |i|
         # get eps for year
         is = eps.select{ |s| s.year.to_i ==  (year - i).to_i  }.first
-      
+
         if is
           revenue = (clean_string(tr[i].text).to_f.round * MILLION).to_s
           income = (clean_string(ni[i].text).to_f.round * MILLION).to_s
@@ -333,7 +334,7 @@ def get_revenue_income_msn
 
   def get_eps_from_msn
     url = "http://investing.money.msn.com/investments/financial-results/?symbol=us%3a#{ticker}"
-    
+
     doc = open_url_or_nil(url)
 
     begin
@@ -341,7 +342,7 @@ def get_revenue_income_msn
       ttm_eps = ttm_eps[13].children[1].text.gsub(",","").to_f
     rescue
     end
-    
+
     if ttm_eps
       update_attributes( :ttm_eps => ttm_eps )
       puts "eps for #{ticker} is #{ttm_eps}"
@@ -451,12 +452,12 @@ def get_revenue_income_msn
   end
 
   # Balace sheet ---------------------------------------------------------------
-  
-  
+
+
   def get_data_from_gurufocus
     # Get data from MSN first, as it is more up to date, reliable, and correct
     # See STO in NOK for example
-    
+
 
     #Acces data page
     url = "http://www.gurufocus.com/financials/#{ticker}"
@@ -465,64 +466,64 @@ def get_revenue_income_msn
       puts "Could not get finantial data from gurufocus.com"
       return false
     end
-    
+
     # Check if year is updated for 2012
     fp = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Fiscal Period" }
     fp = fp.xpath('./td') if fp
-       
+
     if fp.nil?
       puts "--------------------------------------Cannot get info for #{ticker}"
       return false
-    end   
+    end
     # Find last year by counting 'td's up to "TMM"
     years_available = 0 # Some stocks may not have 10 years worth of data
     for i in 1..fp.size
-      if fp[i].nil? || !fp[i].text.match("TTM").nil? 
+      if fp[i].nil? || !fp[i].text.match("TTM").nil?
         break
       end
       years_available = i
     end
-    
+
     puts "Counted #{years_available} years of available data for #{ticker}"
-          
+
     update_year = 1 # Some stocks may not be updated for 2012 yet
     update_year = 0 if fp[years_available].text.last == (YEAR-1).to_s.last
-    
+
     # A boolean to test if current asset values are available
     using_current_data = true
-    
+
     # Scrape data from doc
     # Current Assets
     ca = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Current Assets" }
     if ca
-      ca = ca.xpath('./td') 
+      ca = ca.xpath('./td')
     else
       using_current_data = false
     end
-    
+
     ta = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Assets" }
     ta = ta.xpath('./td') if ta
-    
+
     cl = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Current Liabilities" }
     if cl
-      cl = cl.xpath('./td') 
+      cl = cl.xpath('./td')
     else
       using_current_data = false
     end
-        
+
     tl = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Liabilities" }
     tl = tl.xpath('./td') if tl
-        
+
     # Debt, book value, net tangible assets
     ltd = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Long-Term Debt" }
     ltd = ltd.xpath('./td') if ltd
-    
+
     bv = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Total Equity" }
     bv = bv.xpath('./td') if bv
-    
+
     ocs = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Other Current Assets" }
     ocs = ocs.xpath('./td') if ocs
-    
+
     # Create balance sheet for 10 years
     (1..years_available).each do |i|
       cas = ""
@@ -535,9 +536,9 @@ def get_revenue_income_msn
           ntas = (( clean_string(ca[i].text).to_f - clean_string(ocs[i].text).to_f - clean_string(cl[i].text).to_f ).round * MILLION ).to_s
         else
           ntas = cas
-        end 
+        end
       end
-      
+
       # Some trusts don't have liabilities
       tler = ""
       tler = (clean_string(tl[i].text).to_f.round * MILLION).to_s if tl
@@ -554,25 +555,25 @@ def get_revenue_income_msn
                             :long_term_debt => der,
                             :net_tangible_assets => ntas,
                             :book_value => bver )
-                                     
+
         puts "Got bs data for #{ticker}, year: #{bs.year}, ta = #{bs.total_assets}" if !bs.id.nil?
     end
-    
-    update_attributes( :has_currant_ratio => using_current_data)    
-      
-    #Scrape (from same page): Revenue, net earnings, 
+
+    update_attributes( :has_currant_ratio => using_current_data)
+
+    #Scrape (from same page): Revenue, net earnings,
     # and - diluted EPS, including extra items
     r = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Revenue" }
     r = r.xpath('./td') if r
-    
+
     ni = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Net Income" }
     ni = ni.xpath('./td') if ni
-    
+
     # This is eps from gurufocus, not exactly what we want
     eps = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['title'] == "Earnings per Share (diluted) ($)" }
     eps = eps.xpath('./td') if eps
-    
-    
+
+
     (1..years_available).each do |i|
       cur_year = YEAR - (years_available+1 - i)- update_year
       # does an ep exist for this year?
@@ -585,9 +586,9 @@ def get_revenue_income_msn
                      :revenue => (clean_string(r[i].text).to_f.round * MILLION).to_s,
                      :eps => clean_string(eps[i].text).to_f,
                      :source => url)
-        if !ep.id.nil?            
+        if !ep.id.nil?
           puts "Created new ep for #{ticker}, year: #{ep.year}, rev: #{ep.revenue}, income: #{ep.net_income}, eps: #{ep.eps.to_s}"
-        end  
+        end
       else
         puts "Found ep data for year #{cur_year}"
         if cur_ep && cur_ep.net_income.nil?
@@ -595,122 +596,122 @@ def get_revenue_income_msn
                      :revenue => (clean_string(r[i].text).to_f.round * MILLION).to_s,
                      :source => url)
           puts "Updated ep data for #{ticker}, year: #{cur_ep.year}, rev: #{cur_ep.revenue}, income: #{cur_ep.net_income}"
-        end               
-      end           
+        end
+      end
       #puts "Got eps data for #{ticker}, year: #{YEAR - (11 - i)- update_year}, rev: #{r[i].text}, income: #{ni[i].text}, eps: #{eps[i].text}"
     end
-    
+
     # MSN -- as data source from here on:
-      
+
     # update eps for past 5 years from msn
     # get last five years earnings as 'diluted eps including extra items' from msn
     url = "http://investing.money.msn.com/investments/stock-income-statement/?symbol=US%3a#{ticker}"
-      
+
     doc = open_url_or_nil(url)
-    
+
     if doc.nil?
       puts "ERROR: Could not get data from msn for #{ticker}"
       return
     end
     puts "\n Updating eps data from msn for #{ticker}"
-    
+
     #test if updated for 2013 or not (not neceraly updated to same year as gurufocus)
     last_year = YEAR - 1
     date = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "PeriodEndingDate" } if doc
     if date
-      date = date.xpath('./td') 
+      date = date.xpath('./td')
     else
       puts "cannot get information from msn for #{ticker}"
-      return 
+      return
     end
-    
+
     update_year_msn = 1 #Some stocks may not be updated for 2013 yet
     update_year_msn = 0 if !date[1].text.match(last_year.to_s).nil?
-       
+
     epss = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "DilutedEPS" }  if doc
-    
+
     # if new model data was updated, we need to reload the cache
     self.reload
-    
+
     # Update earnings from msn, if needed and available?
     if latest_eps.year == (last_year-1) && update_year_msn == 0
-     puts "Creating new eps for year #{last_year}" 
-     rev2012 = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "TotalRevenue" }  
+     puts "Creating new eps for year #{last_year}"
+     rev2012 = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "TotalRevenue" }
      rev2012 = rev2012.xpath('./td')[1]
      earn2012 = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "NetIncome" }
-     earn2012 = earn2012.xpath('./td')[1]  
-     eps2012 = epss.xpath('./td')[1]  
+     earn2012 = earn2012.xpath('./td')[1]
+     eps2012 = epss.xpath('./td')[1]
      ep = Ep.create(:stock_id => self.id,
                      :year => last_year,
                      :revenue => (clean_string(rev2012.text).to_f.round * MILLION).to_s,
                      :net_income => (clean_string(earn2012.text).to_f.round * MILLION).to_s,
                      :eps => clean_string(eps2012.text).to_f, # Not that this might be overriden in the next code block
                      :source => url)
-     if !ep.id.nil?            
+     if !ep.id.nil?
        puts "[From msn] Created new ep for #{ticker}, year: #{ep.year}, rev: #{ep.revenue}, income: #{ep.net_income}, eps: #{ep.eps.to_s}"
-     end  
+     end
    end
-     
+
     if epss
       epss = epss.xpath('./td')
-      
+
       #NOte - this updates last 5 year eps regardless of if exist
       (1..5).each do |i|
         if epss[i]
            year = YEAR - i - update_year_msn
-           ep = self.eps.select{ |s| s.year == year }.first   
+           ep = self.eps.select{ |s| s.year == year }.first
            if !ep.nil?
-             ep.update_attributes( :eps => clean_string(epss[i].text).to_f, :source => url) 
+             ep.update_attributes( :eps => clean_string(epss[i].text).to_f, :source => url)
              puts "Updated ep for #{ticker}, year: #{year}, eps: #{clean_string(epss[i].text)}"
-           end  
+           end
         end
       end
     end
-    
+
     # If data is available for most recent past year, add this data for (2) Balance sheets
     if latest_balance_sheet.year == (YEAR-2) && update_year_msn == 0
       url = "http://investing.money.msn.com/investments/stock-balance-sheet/?symbol=us%3A#{ticker}&stmtView=Ann"
       doc = open_url_or_nil(url)
-    
+
       if doc.nil?
         puts "ERROR: Could not get data from msn for #{ticker}"
         return
       end
-      puts "\n Updating balance sheet data from msn for #{ticker} for year #{YEAR-1}" 
-      
+      puts "\n Updating balance sheet data from msn for #{ticker} for year #{YEAR-1}"
+
       cas = cls = tler = der = ntas = bver = ""
-      
-      ca = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "CurrentAssets" }  
+
+      ca = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "CurrentAssets" }
       cas = ca.xpath('./td') if ca
-    
+
       ta = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "TotalAssets" }
       ta = ta.xpath('./td') if ta
-    
+
       cl = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "CurrentLiabilities" }
       if cl
-        cls = cl.xpath('./td') 
+        cls = cl.xpath('./td')
       else
         using_current_data = false
       end
-        
+
       tl = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "TotalLiabilities" }
       tler = tl.xpath('./td') if tl
-        
+
       # Debt, book value, net tangible assets
       ltd = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "LongTermDebt" }
       der = ltd.xpath('./td') if ltd
-    
+
       bv = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "TotalEquity" }
       bver = bv.xpath('./td') if bv
-      
+
       # Get the relavent string, and adjust it for saving in data base ONLY FOR 2012
       cas = (clean_string(cas[1].text).to_f.round * MILLION).to_s if cas != ""
       cls = (clean_string(cls[1].text).to_f.round * MILLION).to_s if cls != ""
-      tler = (clean_string(tler[1].text).to_f.round * MILLION).to_s if tler != ""      
+      tler = (clean_string(tler[1].text).to_f.round * MILLION).to_s if tler != ""
       der = (clean_string(der[1].text).to_f.round * MILLION).to_s if der != ""
       ntas = cas
       bver = (clean_string(bver[1].text).to_f.round * MILLION).to_s if bver != ""
-    
+
       bs = BalanceSheet.create(:stock_id => self.id,
                             :year => YEAR-1,
                             :current_assets => cas,
@@ -720,15 +721,15 @@ def get_revenue_income_msn
                             :long_term_debt => der,
                             :net_tangible_assets => ntas,
                             :book_value => bver )
-                                     
+
       puts "[From msn]Got bs data for #{ticker}, year: #{bs.year}, ta = #{bs.total_assets}" if !bs.id.nil?
-      update_attributes( :has_currant_ratio => using_current_data) 
-         
-    end #end of special update if got 2012 data from msn   
+      update_attributes( :has_currant_ratio => using_current_data)
+
+    end #end of special update if got 2012 data from msn
 
   end #end of methos 'get_data_from_gurofucos' which realy gets data from msn
-  
-  
+
+
 # only gets 5 years back
   def get_bs_from_msn
 
@@ -881,7 +882,7 @@ def get_revenue_income_msn
       diveds = doc.xpath('//div[@id="divhisbotleft"]')
 
       diveds = diveds.children[5].children
-  
+
       diveds.each do |d|
            next if d.nil?
            next if d.text.strip ==  "" # skip empty elements used as spacing by the website
@@ -1013,33 +1014,33 @@ def get_historic_eps(years_back)
 
     # get last five years earnings as 'diluted eps including extra items' from msn
     url = "http://investing.money.msn.com/investments/stock-income-statement/?symbol=US%3a#{ticker}"
-      
+
     # dont download if historic eps data already exists
     # return if self.eps.count >= years_back
-  
+
     doc = open_url_or_nil(url)
-  	start = 1 # how far back to get earnings
-   	year = YEAR - 1
-   	puts "\n Getting earnings records for #{ticker}"
-      
+    start = 1 # how far back to get earnings
+    year = YEAR - 1
+    puts "\n Getting earnings records for #{ticker}"
+
     eps = doc.xpath('//tr').detect{ |tr| tr.xpath('./td').first != nil && tr.xpath('./td').first['id'] == "DilutedEPSIncludingExtraOrdIte" }  if doc
-     
-  	if eps
-      eps = eps.xpath('./td') 
+
+    if eps
+      eps = eps.xpath('./td')
       years_togo_back = min(5,years_back)
       (1..years_togo_back).each do |i|
-	      if eps[i]
- 	   	    ep = Ep.create(:stock_id => self.id,
+        if eps[i]
+          ep = Ep.create(:stock_id => self.id,
                          :year => YEAR - i,
                          :eps => clean_string(eps[i].text).to_f,
                          :source => url)
- 	    		puts "created eps for #{ticker}, year: #{ep.year}, eps: #{ep.eps}"  if !ep.id.nil?
-            	 end
-      	   end
+          puts "created eps for #{ticker}, year: #{ep.year}, eps: #{ep.eps}"  if !ep.id.nil?
+               end
+           end
         end
 
-      # get eps for years 6-10 going back 
-      if years_back > 5 
+      # get eps for years 6-10 going back
+      if years_back > 5
         url = "http://investing.money.msn.com/investments/financial-statements?symbol=US%3a#{ticker}"
 
         doc = open_url_or_nil(url)
@@ -1050,16 +1051,16 @@ def get_historic_eps(years_back)
 
         if !eps.nil? && !eps.empty?
           (5..years_back).each do |i|
-	     if !eps[i].nil?  	 
-         	  ep = Ep.create(:stock_id => self.id,
-                	       	:year => year,
-				:eps =>  clean_string(eps[i].text).to_f,
-                		:source => url)
+       if !eps[i].nil?
+            ep = Ep.create(:stock_id => self.id,
+                          :year => year,
+        :eps =>  clean_string(eps[i].text).to_f,
+                    :source => url)
 
-        	  puts "created eps for #{ticker}, year: #{ep.year}, eps: #{ep.eps}" if !ep.id.nil?
-        	  year = year - 1
-             end	
-      	  end # do i loop
+            puts "created eps for #{ticker}, year: #{ep.year}, eps: #{ep.eps}" if !ep.id.nil?
+            year = year - 1
+             end
+          end # do i loop
         end
 
       end # getting data 6-10 years back
@@ -1069,77 +1070,77 @@ def get_historic_eps(years_back)
  # get number of shares outstanding
 
   def get_numshares
-   
+
     puts "Getting shares outstanding for #{ticker}"
 
     url = "http://investing.money.msn.com/investments/financial-statements?symbol=US%3a#{ticker}"
 
     doc = open_url_or_nil(url)
     return if doc.nil?
-        
+
     sel = doc.css('table[class = " mnytbl"]')
     return if sel.nil? or sel[1].nil?
 
     so = sel[1].css('td[class = "nwrp last"]')
     return if so.nil?
 
-	  year = YEAR - 1
+    year = YEAR - 1
 
     if !so.nil? && !so.empty?
       so.each do |n|
-        if !n.nil?  
+        if !n.nil?
           shares_outstanding = clean_string(n.text)
           next if shares_outstanding.to_i == 0
-         	sharec = Numshare.create(:stock_id => self.id,
-                	       	:year => year,
-				                  :shares => shares_outstanding)
+          sharec = Numshare.create(:stock_id => self.id,
+                          :year => year,
+                          :shares => shares_outstanding)
           puts "created numshare for #{ticker}, year: #{sharec.year}, shares: #{sharec.shares}" if !sharec.id.nil?
-        	year = year - 1
+          year = year - 1
         end
-      end      
+      end
     end
-      
+
   end # method for number of shares
 
 
   def get_dividendsm
-    
+
     markn = "mark5"
     return if self.mark == markn
-    
+
     # Some nasdaq stocks had " " (empty space) at end of name -> correct that:
     if ticker.last == " "
       new_tick = ticker.chop
       update_attributes( :ticker => new_tick )
     end
-    
+
     url = "http://www.nasdaq.com/symbol/#{ticker}/dividend-history"
 
     puts "\n Getting dividends for #{ticker}"
-    
+
     doc = open_url_or_nil(url)
     return if doc.nil?
-    
-    # get table with id: table id="dividendhistoryGrid" 
+
+    # get table with id: table id="dividendhistoryGrid"
     # Better way to do this with css selectors?
     dividends = doc.xpath('//table[@id="dividendhistoryGrid"]/tr').map{ |row| row.xpath('.//td')}
     # Now, dividends[i] is a row
     # dividends[i].children[1] is date of div i
     # dividends[i].children[5] is amount of div i
-      
+
     #.map {|item| item.text.gsub!(/[\r|\n]/,"").strip} }
     # Get the right date format: Date.strptime("2/10/2012", '%m/%d/%Y')
-      
+
     #remove first row
     dividends.delete_at(0)
-      
+
     dividends.each do |d|
       begin
         Date.strptime( d[0].text.gsub!(/[\r|\n]/,"").strip , '%m/%d/%Y')
       rescue
         next # Invalid date, skip to next
       end
-     
+
       div = Dividend.create(:stock_id => self.id,
                             :date => Date.strptime( d[0].text.gsub!(/[\r|\n]/,"").strip , '%m/%d/%Y'),
                             :amount => d[2].text.gsub!(/[\r|\n]/,"").strip.to_f,
@@ -1147,10 +1148,10 @@ def get_historic_eps(years_back)
 
       puts "\n Added dividend record for #{ticker}: date - #{ div.date }, amount: #{div.amount.to_f}" if !div.id.nil?
     end
-      
+
     # Mark stock as handled:
     update_attributes(:mark => markn)
-      
+
   end
 
 end # end module datascraper
