@@ -1,80 +1,78 @@
-ActionController::Routing::Routes.draw do |map|
-  map.resources :searches
-
-  # map.resources :stocks
-
-
-
-  # The priority is based upon order of creation: first created -> highest priority.
+Graham::Application.routes.draw do
+  # The priority is based upon order of creation:
+  # first created -> highest priority.
 
   # Sample of regular route:
-  #   map.connect 'products/:id', :controller => 'catalog', :action => 'view'
+  #   match 'products/:id' => 'catalog#view'
   # Keep in mind you can assign values other than :controller and :action
 
   # Sample of named route:
-  #   map.purchase 'products/:id/purchase', :controller => 'catalog', :action => 'purchase'
+  #   match 'products/:id/purchase' => 'catalog#purchase', :as => :purchase
   # This route can be invoked with purchase_url(:id => product.id)
 
   # Sample resource route (maps HTTP verbs to controller actions automatically):
-  #   map.resources :products
+  #   resources :products
 
   # Sample resource route with options:
-  #   map.resources :products, :member => { :short => :get, :toggle => :post }, :collection => { :sold => :get }
+  #   resources :products do
+  #     member do
+  #       get 'short'
+  #       post 'toggle'
+  #     end
+  #
+  #     collection do
+  #       get 'sold'
+  #     end
+  #   end
 
   # Sample resource route with sub-resources:
-  #   map.resources :products, :has_many => [ :comments, :sales ], :has_one => :seller
+  #   resources :products do
+  #     resources :comments, :sales
+  #     resource :seller
+  #   end
 
   # Sample resource route with more complex sub-resources
-  #   map.resources :products do |products|
-  #     products.resources :comments
-  #     products.resources :sales, :collection => { :recent => :get }
+  #   resources :products do
+  #     resources :comments
+  #     resources :sales do
+  #       get 'recent', :on => :collection
+  #     end
   #   end
 
   # Sample resource route within a namespace:
-  #   map.namespace :admin do |admin|
-  #     # Directs /admin/products/* to Admin::ProductsController (app/controllers/admin/products_controller.rb)
-  #     admin.resources :products
+  #   namespace :admin do
+  #     # Directs /admin/products/* to Admin::ProductsController
+  #     # (app/controllers/admin/products_controller.rb)
+  #     resources :products
   #   end
 
-  # You can have the root of your site routed with map.root -- just remember to delete public/index.html.
-  # map.root :controller => "welcome"
+  # You can have the root of your site routed with "root"
+  # just remember to delete public/index.html.
+  # root :to => 'welcome#index'
 
   # See how all your routes lay out with "rake routes"
-  map.resources :portfolios,
+
+  # This is a legacy wild controller route that's not recommended for RESTful applications.
+  # Note: This route will make all actions in every controller accessible via GET requests.
+  # match ':controller(/:action(/:id))(.:format)'
+  resources :searches,
+  :portfolios,
   :users,
   :transactions,
   :stocks,
-  :stocks,
   :splits,
   :eps
-  map.defensive_stocks 'defensive_stocks',
-  :controller => 'stocks',
-  :action => 'defensive_stocks'
-  map.cheap_stocks 'cheap_stocks',
-  :controller => 'stocks',
-  :action => 'cheap_stocks'
-  map.cheap_stocks 'dear_stocks',
-  :controller => 'stocks',
-  :action => 'dear_stocks'
-  map.defensive_buys 'defensive_buys',
-  :controller => 'stocks',
-  :action => 'defensive_buys'
-  map.pots 'pots',
-  :controller => 'stocks',
-  :action => 'pots'
-  map.cheap_profitables 'cheap_profitables',
-  :controller => 'stocks',
-  :action => 'cheap_profitables'
-  map.new_buy_trnasaction ':portfolio_id/new_transaction/:type',
-  :controller => 'transactions',
-  :action => 'new'
 
 
+  match "defensive_stocks", :to => "stocks#defensive_stocks"
+  match 'cheap_stocks', :to => "stocks#cheap_stocks"
+  match "dear_stocks", :to => "stocks#dear_stocks"
+  match 'defensive_buys', :to => "stocks#defensive_buys"
+  match "pots", :to => "stocks#pots"
+  match "cheap_profitables", :to => "stocks#cheap_profitables"
+  match ":portfolio_id/new_transaction/:type", :to => "transactions#new"
 
+  #match ":controller(/:action(/:id))"
+  #match ":controller(/:action(/:id))(.:format)"
 
-  # Install the default routes as the lowest priority.
-  # Note: These default routes make all actions in every controller accessible via GET requests. You should
-  # consider removing or commenting them out if you're using named routes and resources.
-  map.connect ':controller/:action/:id'
-  map.connect ':controller/:action/:id.:format'
 end
